agent/requirements.txt
websockets>=12.0
aiohttp>=3.9
pyyaml>=6.0
chromadb>=0.5
anthropic>=0.30
openai>=1.30
google-generativeai>=0.7
pydantic>=2.0
